moved code that update the dsl file from command to proper wizard
