+++ framework/xdsml_framework/plugins/org.eclipse.gemoc.xdsmlframework.extensions.sirius/META-INF/MANIFEST.MF
@@ -28,9 +28,10 @@ Require-Bundle: org.eclipse.gemoc.xdsmlframework.api,
  org.eclipse.gemoc.xdsmlframework.ui.utils,
  org.eclipse.gemoc.commons.eclipse.ui,
  org.eclipse.swt,
- org.eclipse.ui.workbench
+ org.eclipse.ui.workbench,
+ org.eclipse.gemoc.dsl.model
 Bundle-ActivationPolicy: lazy
-Bundle-RequiredExecutionEnvironment: JavaSE-1.7
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Export-Package: org.eclipse.gemoc.xdsmlframework.extensions.sirius,
  org.eclipse.gemoc.xdsmlframework.extensions.sirius.command,
  org.eclipse.gemoc.xdsmlframework.extensions.sirius.wizards,
