+++ framework/xdsml_framework/plugins/org.eclipse.gemoc.xdsmlframework.extensions.sirius/src/main/java/org/eclipse/gemoc/xdsmlframework/extensions/sirius/wizards/NewGemocSiriusProjectWizard.java
@@ -10,8 +10,12 @@
  *******************************************************************************/
 package org.eclipse.gemoc.xdsmlframework.extensions.sirius.wizards;
 
+import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
+import java.util.Collections;
+import java.util.Optional;
 
+import org.eclipse.core.resources.IFile;
 import org.eclipse.core.resources.IProject;
 import org.eclipse.core.resources.IProjectDescription;
 import org.eclipse.core.resources.IResource;
@@ -20,11 +24,21 @@ import org.eclipse.core.resources.IWorkspaceRunnable;
 import org.eclipse.core.resources.ResourcesPlugin;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IProgressMonitor;
+import org.eclipse.core.runtime.OperationCanceledException;
 import org.eclipse.core.runtime.Path;
+import org.eclipse.core.runtime.jobs.Job;
+import org.eclipse.emf.common.util.URI;
+import org.eclipse.emf.ecore.resource.Resource;
+import org.eclipse.emf.ecore.resource.impl.ResourceSetImpl;
 import org.eclipse.emf.edit.ui.provider.ExtendedImageRegistry;
+import org.eclipse.gemoc.commons.eclipse.core.resources.FileFinderVisitor;
 import org.eclipse.gemoc.commons.eclipse.pde.wizards.pages.pde.AbstractNewProjectWizardWithTemplates;
 import org.eclipse.gemoc.commons.eclipse.pde.wizards.pages.pde.ui.IProjectContentWizard;
 import org.eclipse.gemoc.commons.eclipse.pde.wizards.pages.pde.ui.ProjectTemplateApplicationOperation;
+import org.eclipse.gemoc.dsl.Dsl;
+import org.eclipse.gemoc.dsl.DslFactory;
+import org.eclipse.gemoc.dsl.DslPackage;
+import org.eclipse.gemoc.dsl.Entry;
 import org.eclipse.gemoc.xdsmlframework.extensions.sirius.Activator;
 import org.eclipse.gemoc.xdsmlframework.extensions.sirius.wizards.pages.NewGemocSiriusProjectMainWizardPage;
 import org.eclipse.gemoc.xdsmlframework.extensions.sirius.wizards.pages.NewGemocSiriusProjectWizardFields;
@@ -117,8 +131,7 @@ public class NewGemocSiriusProjectWizard extends AbstractNewProjectWizardWithTem
 									newOdesignPage.getEncoding(), 
 									getContainer());
 					
-					// launch the template
-					
+					// launch the template is selected
 					IProjectContentWizard contentWizard = templateSelectionPage.getSelectedWizard();
 					try {
 						getContainer().run(false, true, new ProjectTemplateApplicationOperation(context, project, contentWizard));
@@ -128,7 +141,11 @@ public class NewGemocSiriusProjectWizard extends AbstractNewProjectWizardWithTem
 						Activator.logErrorMessage(e.getMessage(), e);
 					}
 					
-					//setClassPath(project, monitor);
+					project.refreshLocal(IResource.DEPTH_INFINITE, monitor);
+					
+					waitForAutoBuild();
+					updateDsl(project);
+					
 					project.refreshLocal(IResource.DEPTH_INFINITE, monitor);
 				}
 			};
@@ -141,7 +158,61 @@ public class NewGemocSiriusProjectWizard extends AbstractNewProjectWizardWithTem
 		return true;
 	}
 	
+	/**
+	 * update the dsl file with the indication of the sirius project 
+	 * @param siriusProject
+	 */
+	protected void updateDsl(IProject siriusProject){
+		
+		FileFinderVisitor odesignProjectVisitor = new FileFinderVisitor("odesign");
+		try {
+			siriusProject.accept(odesignProjectVisitor);
+			IFile odesignIFile = odesignProjectVisitor.getFile();
+			if (odesignIFile != null) {
+				// the dsl file that will be updated
+				Resource res = (new ResourceSetImpl()).getResource(URI.createURI(context.dslFilePath), true);
+				Dsl dsl = (Dsl) res.getContents().get(0);
+				
+				Optional<Entry> sirius = dsl.getEntries()
+					.stream()
+					.filter(entry -> entry.getKey().equals("sirius"))
+					.findFirst();
+				if(sirius.isPresent()) {
+					sirius.get().setValue(odesignIFile.getFullPath().toOSString());
+				}
+				else {
+					Entry siriusEntry = ((DslFactory)DslPackage.eINSTANCE.getEFactoryInstance()).createEntry();
+					siriusEntry.setKey("sirius");
+					siriusEntry.setValue(odesignIFile.getFullPath().toOSString());
+					dsl.getEntries().add(siriusEntry);
+				}
+				try {
+					res.save(Collections.emptyMap());
+				} catch (IOException e) {
+					Activator.logErrorMessage(e.getMessage(), e);
+				}
+			}
+		} catch (CoreException e) {
+			Activator.logErrorMessage(e.getMessage(), e);
+		}
+	}
 	
+	/**
+	 * wait that eclipse worskace has finish all its background builds
+	 */
+	protected void waitForAutoBuild() {
+		boolean wasInterrupted = false;
+		do {
+			try {
+				Job.getJobManager().join(ResourcesPlugin.FAMILY_AUTO_BUILD,	null);
+				wasInterrupted = false;
+			} catch (OperationCanceledException e) {
+				Activator.logWarnMessage(e.getMessage(), e);
+			} catch (InterruptedException e) {
+				wasInterrupted = true;
+			}
+		} while (wasInterrupted);
+	}
 	@Override
 	public boolean isHelpAvailable() {
 		return true;
